+++ bcdi/preprocessing/analysis.py
@@ -489,7 +489,6 @@ class Analysis(ABC):
                 tilt_values=None,
                 logger=self.logger,
                 plot_fit=False,
-                frames_pattern=self.parameters["frames_pattern"],
             )
             self.q_bragg = np.array(
                 [
@@ -664,7 +663,6 @@ class Analysis(ABC):
             peak_method=self.parameters["centering_method"]["reciprocal_space"],
             tilt_values=self.setup.tilt_angles,
             savedir=self.setup.detector.savedir,
-            frames_pattern=self.parameters.get("frames_pattern"),
             user_defined_peak=self.parameters["bragg_peak"],
             logger=self.logger,
             plot_fit=True,
@@ -1175,7 +1173,7 @@ class FirstDataLoading(PreprocessingLoader):
         return bu.load_bcdi_data(
             scan_number=self.scan_nb,
             setup=self.setup,
-            frames_pattern=self.parameters["frames_pattern"],
+            frames_pattern=self.parameters.get("frames_pattern"),
             bin_during_loading=self.parameters["bin_during_loading"],
             flatfield=util.load_flatfield(self.parameters["flatfield_file"]),
             hotpixels=util.load_hotpixels(self.parameters["hotpixels_file"]),